Modified to work with data in the form of a row vector or a column vector.
Modified to work with data in the form of a row vector or a column vector.

Now checks the orientation of the data, and if the data is in the form of a row vector, ithe data is first converted into a column vector before filtering, and then converted back to a row vector after filtering.