Fixed bug that was not filtering the data if the data was in the form of a column vector.
Fixed bug that was not filtering the data if the data was in the form of a column vector.

Now checks the orientation of the data, and if the data is a row vector, it is converted into a column vector before filtering, and then converted back into a row vector after filtering.